HCK-7336: move an alter tag helper to separate file
