+++ forward_engineering/api.js
@@ -3,7 +3,6 @@
 const _ = require('lodash');
 const { commentDropStatements } = require('./helpers/commentDropStatements');
 const { DROP_STATEMENTS } = require('./helpers/constants');
-const { buildContainerLevelAlterScript } = require('./alterScript/alterScriptBuilder');
 
 module.exports = {
 	generateScript(data, logger, callback, app) {
@@ -35,10 +34,8 @@ module.exports = {
 
 	generateContainerScript(data, logger, callback, app) {
 		try {
-			const ddlProvider = require('./ddlProvider')(_, data.options, app);
-			const script = buildContainerLevelAlterScript({ data, app, ddlProvider });
-
-			callback(null, script);
+			data.jsonSchema = data.collections[0];
+			this.generateScript(data, logger, callback, app);
 		} catch (error) {
 			logger.log('error', { message: error.message, stack: error.stack }, 'Snowflake Forward-Engineering Error');
